chore: rename core path to make way for todo
