--- src/cdf/core/__init__.py
@@ -1,3 +1,0 @@
-from cdf.core.project import DataPackage, Project
-
-__all__ = ["Project", "DataPackage"]
