+++ src/cdf/legacy/__init__.py
@@ -1,0 +1,3 @@
+from cdf.legacy.project import DataPackage, Project
+
+__all__ = ["Project", "DataPackage"]
