+++ src/cdf/legacy/adapter/__init__.py
@@ -1,0 +1,21 @@
+"""Adapters and factories for the core components of the CDF framework."""
+
+import cdf.legacy.adapter.ingest as ingest
+import cdf.legacy.adapter.state as state
+import cdf.legacy.adapter.test as test
+import cdf.legacy.adapter.transform as transform
+from cdf.legacy.adapter.ingest import ingest_adapter_factory
+from cdf.legacy.adapter.state import state_backend_factory
+from cdf.legacy.adapter.test import test_adapter_factory
+from cdf.legacy.adapter.transform import transform_adapter_factory
+
+__all__ = [
+    "ingest",
+    "state",
+    "test",
+    "transform",
+    "ingest_adapter_factory",
+    "state_backend_factory",
+    "test_adapter_factory",
+    "transform_adapter_factory",
+]
